pyproject.toml
[tool.poetry]
name = "learning-platform-api"
version = "0.1.0"
description = ""
authors = ["ncgordeev <ncgordeev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
django = "^5.1"
pillow = "^10.4.0"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
ipython = "^8.27.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.1"
drf-yasg = "^1.21.7"
stripe = "^10.9.0"
celery = "^5.4.0"
redis = "^5.0.8"
django-celery-beat = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
